Fix EventTrait / Event
